backend/requirements.txt
Flask
boto3
python-dotenv
flask-cors
moto
pytest
